move FieldDesc to the "desc" package
The `io.quarkus.gizmo2.desc` package contains all other descriptor
types added by Gizmo (such as `MemberDesc` or `MethodDesc`), so
it's only logical it also contains `FieldDesc`.
