+++ src/main/java/io/quarkus/gizmo2/Expr.java
@@ -2,6 +2,7 @@ package io.quarkus.gizmo2;
 
 import java.lang.constant.ClassDesc;
 
+import io.quarkus.gizmo2.desc.FieldDesc;
 import io.quarkus.gizmo2.impl.GizmoImpl;
 import io.quarkus.gizmo2.impl.Item;
