+++ src/main/java/io/quarkus/gizmo2/desc/FieldDesc.java
@@ -1,9 +1,9 @@
-package io.quarkus.gizmo2;
+package io.quarkus.gizmo2.desc;
 
 import java.lang.constant.ClassDesc;
 import java.util.Objects;
 
-import io.quarkus.gizmo2.desc.MemberDesc;
+import io.quarkus.gizmo2.SimpleTyped;
 import io.quarkus.gizmo2.impl.FieldDescImpl;
 import io.quarkus.gizmo2.impl.GizmoImpl;
 import io.quarkus.gizmo2.impl.Util;
