+++ src/main/java/io/quarkus/gizmo2/impl/GizmoImpl.java
@@ -9,14 +9,14 @@ import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
-import io.quarkus.gizmo2.desc.ClassMethodDesc;
 import io.quarkus.gizmo2.ClassOutput;
-import io.quarkus.gizmo2.FieldDesc;
 import io.quarkus.gizmo2.Gizmo;
-import io.quarkus.gizmo2.desc.InterfaceMethodDesc;
 import io.quarkus.gizmo2.StaticFieldVar;
 import io.quarkus.gizmo2.creator.ClassCreator;
 import io.quarkus.gizmo2.creator.InterfaceCreator;
+import io.quarkus.gizmo2.desc.ClassMethodDesc;
+import io.quarkus.gizmo2.desc.FieldDesc;
+import io.quarkus.gizmo2.desc.InterfaceMethodDesc;
 import io.quarkus.gizmo2.impl.constant.ArrayVarHandleConstant;
 import io.quarkus.gizmo2.impl.constant.ByteConstant;
 import io.quarkus.gizmo2.impl.constant.CharConstant;
