+++ src/main/java/io/quarkus/gizmo2/impl/StaticFieldVarImpl.java
@@ -9,8 +9,8 @@ import java.util.function.BiFunction;
 
 import io.github.dmlloyd.classfile.CodeBuilder;
 import io.quarkus.gizmo2.AccessMode;
-import io.quarkus.gizmo2.FieldDesc;
 import io.quarkus.gizmo2.StaticFieldVar;
+import io.quarkus.gizmo2.desc.FieldDesc;
 import io.quarkus.gizmo2.impl.constant.ConstantImpl;
 
 public final class StaticFieldVarImpl extends LValueExprImpl implements StaticFieldVar {
