+++ src/main/java/io/quarkus/gizmo2/impl/constant/ConstantImpl.java
@@ -14,9 +14,9 @@ import java.util.Objects;
 import io.github.dmlloyd.classfile.CodeBuilder;
 import io.github.dmlloyd.classfile.TypeKind;
 import io.quarkus.gizmo2.Constant;
-import io.quarkus.gizmo2.FieldDesc;
 import io.quarkus.gizmo2.InvokeKind;
 import io.quarkus.gizmo2.desc.ConstructorDesc;
+import io.quarkus.gizmo2.desc.FieldDesc;
 import io.quarkus.gizmo2.desc.MethodDesc;
 import io.quarkus.gizmo2.impl.BlockCreatorImpl;
 import io.quarkus.gizmo2.impl.Item;
